requirements.txt
black==20.8b1
torch==1.7.1
tqdm==4.58.0
numpy==1.20.1
pytorch-lightning==1.1.7
wandb==0.10.20
mypy==0.812
hydra-core==1.0.6
omegaconf==2.0.6
